cambios en backend por javi

+++ backend/src/main/java/com/greenfieldlibrary/backend/Persistence/Lendings.java
@@ -15,64 +15,64 @@ import jakarta.persistence.Table;
 @Table(name = "lendings")
 public class Lendings {
 
-@Id
-@GeneratedValue(strategy = GenerationType.IDENTITY)
-private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
 
-@ManyToOne
-@JoinColumn(name = "id_books") 
-private Books book;
+    @ManyToOne
+    @JoinColumn(name = "id_books")
+    private Books book;
 
-@ManyToOne
-@JoinColumn(name = "id_member")
-private Members member;
+    @ManyToOne
+    @JoinColumn(name = "id_member")
+    private Members member;
 
-@Column(name = "data_lending")
-private Date dataLending;
+    @Column(name = "data_lending")
+    private Date dataLending;
 
-@Column(name = "data_return")
-private Date dataReturn;
+    @Column(name = "data_return")
+    private Date dataReturn;
 
-public Lendings() {
-}
+    public Lendings() {
+    }
 
-public Long getId() {
-return id;
-}
+    public Integer getId() {
+        return id;
+    }
 
-public void setId(Long id) {
-this.id = id;
-}
+    public void setId(Integer id) {
+        this.id = id;
+    }
 
-public Books getBook() {
-return book;
-}
+    public Books getBook() {
+        return book;
+    }
 
-public void setBook(Books book) {
-this.book = book;
-}
+    public void setBook(Books book) {
+        this.book = book;
+    }
 
-public Members getMember() {
-return member;
-}
+    public Members getMember() {
+        return member;
+    }
 
-public void setMember(Members member) {
-this.member = member;
-}
+    public void setMember(Members member) {
+        this.member = member;
+    }
 
-public Date getDataLending() {
-return dataLending;
-}
+    public Date getDataLending() {
+        return dataLending;
+    }
 
-public void setDataLending(Date dataLending) {
-this.dataLending = dataLending;
-}
+    public void setDataLending(Date dataLending) {
+        this.dataLending = dataLending;
+    }
 
-public Date getDataReturn() {
-return dataReturn;
-}
+    public Date getDataReturn() {
+        return dataReturn;
+    }
 
-public void setDataReturn(Date dataReturn) {
-this.dataReturn = dataReturn;
-}
+    public void setDataReturn(Date dataReturn) {
+        this.dataReturn = dataReturn;
+    }
 }
